+++ app/background/setting/client.js
@@ -8,4 +8,6 @@ export const clientStub = ipcRendererInjector => makeClient(ipcRendererInjector,
   'getCustomLocale',
   'setCustomLocale',
   'getLatestRelease',
+  'getDnsFeeSpeed',
+  'setDnsFeeSpeed',
 ]);